visually fixes collisions
i also tried to forward predict rendering positions and modify the
speeds slightly to force balls to hit exactly at a multiple of the
update delta, but the it appears the delta is off by a little each time,
so the estimation is off
oh well. this hacky bit works too.